Remove svelte-data-sync and references to it
