+++ projects/plugins/boost/app/assets/src/js/features/critical-css/lib/stores/critical-css-state-types.ts
@@ -1,5 +1,5 @@
+import { JSONSchema } from '$lib/utils/json-types';
 import z from 'zod';
-import { JSONSchema } from '$lib/stores/data-sync-client';
 
 const CriticalCssErrorType = z.enum( [
 	'SuccessTargetError',
