+++ projects/plugins/boost/app/assets/src/js/lib/api/api-error.ts
@@ -1,7 +1,6 @@
 import { __, sprintf } from '@wordpress/i18n';
-import { JSONObject } from '../stores/data-sync-client';
 import { castToString } from '../utils/cast-to-string';
-import { isJsonObject } from '../utils/json-types';
+import { JSONObject, isJsonObject } from '../utils/json-types';
 
 /**
  * Special error subclass returned by API Calls with extra
