+++ projects/plugins/boost/rollup-tsconfig.json
@@ -4,7 +4,6 @@
 		"../../js-packages/components/**/*",
 		"../../js-packages/image-guide/**/*",
 		"../../js-packages/boost-score-api/**/*",
-		"../../js-packages/svelte-data-sync-client/**/*",
 		"../../js-packages/react-data-sync-client/**/*",
 		"app/assets/src/js/**/*",
 		"app/modules/image-guide/src/**/*"
